+++ src/data/resultReducers.ts
@@ -50,7 +50,13 @@ export function createStoreReducer(
     });
     // TODO add info about networkStatus
 
-    const nextResult = resultReducer(currentResult, action, variables); // action should include operation name
+    let nextResult;
+    try {
+      nextResult = resultReducer(currentResult, action, variables); // action should include operation name
+    } catch (err) {
+      console.warn('Error in a resultReducer, ignoring and moving on.  ', err);
+      throw err;
+    }
 
     if (currentResult !== nextResult) {
       return writeResultToStore({